+++ ondemand/worker/__init__.py
@@ -30,4 +30,6 @@ from .reporter import WorkflowReporter
 from .logging import setup_logging, get_collected_logs, get_and_clear_logs
 from .activity_reporter import report, ActivityReporter
 
-__all__ = ["OndemandWorker", "WorkflowReporter", "setup_logging", "get_collected_logs", "get_and_clear_logs", "report", "ActivityReporter"]
+from ondemand.worker.input import WorkflowInput
+
+__all__ = ["OndemandWorker", "WorkflowInput", "WorkflowReporter", "setup_logging", "get_collected_logs", "get_and_clear_logs", "report", "ActivityReporter"]